--- flaskcode/newtemplate/dashboard.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    <script src="{{ url_for('static', filename='js/scripts.js') }}" defer></script>
-    <title>Senu - Dashboard</title>
-</head>
-<body>
-    <nav>
-        <div class="navbar-brand">Senu</div>
-        <div class="navbar-nav">
-            <a href="{{ url_for('index') }}">Home</a>
-            <a href="{{ url_for('dashboard') }}">Dashboard</a>
-            <a href="{{ url_for('feedback') }}">Feedback</a>
-            <a href="{{ url_for('resources') }}">Resources</a>
-            <a href="{{ url_for('signin') }}">Sign In</a>
-        </div>
-    </nav>
-
-    <div class="container">
-        <h1>Dashboard</h1>
-        <p>Analyze your pitch deck page by page and receive valuable feedback.</p>
-        <div class="analysis-box">
-            <h2>Analysis Section</h2>
-            <p>Feedback and score for the current section will be displayed here.</p>
-        </div>
-    </div>
-</body>
-</html>
