--- flaskcode/newtemplate/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Pitch Deck Upload</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    <script src="{{ url_for('static', filename='js/script.js') }}" defer></script>
-</head>
-<body>
-    <header>
-        <nav>
-            <div class="logo">Senu</div>
-            <ul class="nav-links">
-                <li><a href="/">Home</a></li>
-                <li><a href="/entrepreneur">Entrepreneur Dashboard</a></li>
-                <li><a href="/feedback">Feedback</a></li>
-                <li><a href="/resources">Resources</a></li>
-                <li><a href="/signin">Sign In</a></li>
-            </ul>
-        </nav>
-    </header>
-    <main>
-        <section class="hero">
-            <h2>Upload Your Pitch Deck</h2>
-            <p>Submit your pitch deck for review and analysis.</p>
-            <form id="pitchDeckForm" action="/upload" method="post" enctype="multipart/form-data">
-                <input type="email" name="email" placeholder="Your Email" required>
-                <input type="file" name="pitch_deck" accept=".pdf" required>
-                <button type="submit">Upload Pitch Deck</button>
-            </form>
-        </section>
-    </main>
-</body>
-</html>
