--- flaskcode/newtemplate/static/css/styles.css
@@ -1,256 +1,0 @@
-/* General Styles */
-body {
-    font-family: 'Arial', sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #f4f4f4;
-    color: #333;
-}
-
-header {
-    background-color: #4CAF50;
-    color: #fff;
-    padding: 15px 0;
-}
-
-nav {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-
-.logo {
-    font-size: 24px;
-    font-weight: bold;
-}
-
-.nav-links {
-    list-style: none;
-    display: flex;
-    gap: 20px;
-}
-
-.nav-links li {
-    margin: 0;
-}
-
-.nav-links a {
-    color: #fff;
-    text-decoration: none;
-    font-size: 16px;
-    transition: color 0.3s;
-}
-
-.nav-links a:hover {
-    color: #ddd;
-}
-
-/* Hero Section */
-.hero {
-    background: #4CAF50;
-    color: #fff;
-    padding: 50px 20px;
-    text-align: center;
-}
-
-.hero h2 {
-    font-size: 36px;
-    margin-bottom: 20px;
-}
-
-.hero p {
-    font-size: 18px;
-    margin-bottom: 20px;
-}
-
-.hero form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 10px;
-}
-
-.hero form input[type="email"], 
-.hero form input[type="file"] {
-    padding: 10px;
-    font-size: 16px;
-    border: none;
-    border-radius: 5px;
-    width: 100%;
-    max-width: 300px;
-}
-
-.hero form button {
-    padding: 10px 20px;
-    font-size: 16px;
-    border: none;
-    border-radius: 5px;
-    background-color: #fff;
-    color: #4CAF50;
-    cursor: pointer;
-    transition: background-color 0.3s, color 0.3s;
-}
-
-.hero form button:hover {
-    background-color: #4CAF50;
-    color: #fff;
-}
-
-/* Sign In Section */
-.signin {
-    padding: 50px 20px;
-    text-align: center;
-}
-
-.form-container {
-    background: #fff;
-    padding: 30px;
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    max-width: 400px;
-    margin: 0 auto;
-}
-
-.form-container h2 {
-    margin-bottom: 20px;
-}
-
-.form-container form {
-    display: flex;
-    flex-direction: column;
-    gap: 15px;
-}
-
-.form-container form input[type="email"], 
-.form-container form input[type="password"] {
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
-
-.form-container form button {
-    padding: 10px;
-    font-size: 16px;
-    border: none;
-    border-radius: 5px;
-    background-color: #4CAF50;
-    color: #fff;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.form-container form button:hover {
-    background-color: #45a049;
-}
-
-/* Dashboard Section */
-.dashboard {
-    padding: 50px 20px;
-    text-align: center;
-}
-
-.dashboard h2 {
-    font-size: 28px;
-    margin-bottom: 20px;
-}
-
-.wizard {
-    background: #fff;
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    max-width: 800px;
-    margin: 0 auto;
-    text-align: left;
-}
-
-.wizard h3 {
-    font-size: 24px;
-    margin-bottom: 20px;
-}
-
-.progress {
-    width: 100%;
-    background-color: #f3f3f3;
-    border-radius: 5px;
-    overflow: hidden;
-    margin-bottom: 20px;
-}
-
-.progress-bar {
-    height: 20px;
-    background-color: #4CAF50;
-    width: 0;
-    text-align: center;
-    color: #fff;
-    transition: width 0.3s;
-}
-
-.analysis-boxes {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-}
-
-.analysis-boxes p {
-    background: #f9f9f9;
-    padding: 10px;
-    border-radius: 5px;
-    border: 1px solid #ddd;
-}
-
-/* Feedback Section */
-.feedback {
-    padding: 50px 20px;
-    text-align: center;
-}
-
-.feedback h2 {
-    font-size: 28px;
-    margin-bottom: 20px;
-}
-
-.score-details {
-    background: #fff;
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    max-width: 800px;
-    margin: 0 auto;
-    text-align: left;
-}
-
-/* Resources Section */
-.resources {
-    padding: 50px 20px;
-    text-align: center;
-}
-
-.resources h2 {
-    font-size: 28px;
-    margin-bottom: 20px;
-}
-
-.resource-list {
-    background: #fff;
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    max-width: 800px;
-    margin: 0 auto;
-    text-align: left;
-}
-
-footer {
-    background-color: #222;
-    color: #fff;
-    text-align: center;
-    padding: 10px 0;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-}
